Use reference count for certain objects
Since we are using pointers and sharing this with Perl, we need to keep
track of the reference count for certain kinds of objects that use
shared data on the C++ side. In particular, this is needed for
`kiwi::Variable` and `kiwi::Constraint`. This prevents a double free
error.

This requires a patch to the C++ code to access the necessary private
data. See <https://github.com/project-renard/p5-Alien-Kiwisolver/pull/2>.